Access phonegap API
This file allow user to access phonegap geolocation and notification API as angularjs factory. Easy to implement and plugin.
User just need to add the reference of this file before creating module and then add "PhonegapServices" as reference to your module.

Use factory "pgGeolocationFactory" and "pgNotificationFactory" with alert and confirm only at this stage.